Update LIA - HAPPINESS INDEX - CODE - VO.py
This code creates a line graph that compares countries happiness (ladder score) and freedom levels. it then plots both happiness and freedom as two seperate lines on the same chart. The grid, labels, and legend make the graph easy to read. This allows you to see how freedom tends to change as happiness decreases.